refactor: rule plugin
