+++ src/plugins/spy-middleware.ts
@@ -2,7 +2,7 @@ import type { Options as HttpProxyMiddlewareOptions } from 'http-proxy-middlewar
 import { createProxyMiddleware, responseInterceptor } from 'http-proxy-middleware';
 import { env } from '@/environment';
 import { stringLogger } from '@/logger';
-import { SpyPlugin } from './plugins';
+import { SpyRulePlugin } from './spy-rule-plugin';
 /**
  * Configure proxy middleware
  */
@@ -11,7 +11,7 @@ export const spyMiddleware = createProxyMiddleware({
   changeOrigin: true, // for vhosted sites, changes host header to match to target's host
   selfHandleResponse: true, // manually call res.end(); IMPORTANT: res.end() is called internally by responseInterceptor()
   onProxyRes: responseInterceptor(async (responseBuffer, proxyRes, req, res) => {
-    return await new SpyPlugin(env.srpUpstreamUrl).handleResponse({ responseBuffer, proxyRes, req, res });
+    return await new SpyRulePlugin(env.srpUpstreamUrl).handleResponse({ responseBuffer, proxyRes, req, res });
   }),
   /**
    * TODO: Validate input later
